diabetic-retinopathy/requirements.txt
ipython==8.7.0
matplotlib==3.0.3
numpy==1.24.0
opencv_python==4.1.0.25
pandas==1.0.5
paramiko==2.12.0
Pillow==9.3.0
pydicom==2.3.1
resnet==0.1
scikit_learn==1.2.0
tensorflow==2.11.0
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
~umpy==1.16.4
